+++ src/App.css
@@ -1,5 +1,5 @@
 *{
-  background-color: #291A54;
+  background-color: #071433;
   color: #E3DEB8;
 }
 .App {
