+++ src/components/App.jsx
@@ -1,4 +1,4 @@
-import './App.css';
+import '../App.css';
 import { BrowserRouter as Router, Switch, Route } from 'react-router-dom';
 import React from 'react';
 import Home from './Home';
